# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: The Jobs MV Phase 1 API
servers:
  - url: /api
paths:
...
  /admin/overview:
    get:
      operationId: getAdminOverview
      security:
        - ClerkCookieAuth: []
      responses:
        "200":
          description: Administration overview
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AdminOverview"
        "401":
          $ref: "#/components/responses/Error"
        "403":
          $ref: "#/components/responses/Error"
  /admin/function-check:
    get:
      operationId: getAdminFunctionCheck
      security:
        - ClerkCookieAuth: []
      responses:
        "200":
          description: Sanitized service checks
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AdminFunctionCheck"
        "401":
          $ref: "#/components/responses/Error"
        "403":
          $ref: "#/components/responses/Error"
  /admin/employers:
    get:
      operationId: listAdminEmployers
      security:
        - ClerkCookieAuth: []
      parameters:
        - $ref: "#/components/parameters/Limit"
        - $ref: "#/components/parameters/Offset"
      responses:
        "200":
          description: Employers
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AdminEmployersPage"
        "401":
          $ref: "#/components/responses/Error"
        "403":
          $ref: "#/components/responses/Error"
  /admin/employers/{id}/verification:
    patch:
      operationId: updateAdminEmployerVerification
      parameters:
        - $ref: "#/components/parameters/Id"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/EmployerVerificationUpdate"
      security:
        - ClerkCookieAuth: []
      responses:
        "200":
          description: Updated employer
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AdminEmployer"
        "401":
          $ref: "#/components/responses/Error"
        "403":
          $ref: "#/components/responses/Error"
  /admin/jobs:
    get:
      operationId: listAdminJobs
      security:
        - ClerkCookieAuth: []
      parameters:
        - $ref: "#/components/parameters/Limit"
        - $ref: "#/components/parameters/Offset"
      responses:
        "200":
          description: Jobs
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AdminJobsPage"
        "401":
          $ref: "#/components/responses/Error"
        "403":
          $ref: "#/components/responses/Error"
  /admin/jobs/{id}/close:
    post:
      operationId: closeAdminJob
      security:
        - ClerkCookieAuth: []
      parameters:
        - $ref: "#/components/parameters/Id"
      responses:
        "200":
          description: Closed job
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AdminJob"
        "401":
          $ref: "#/components/responses/Error"
        "403":
          $ref: "#/components/responses/Error"
  /admin/jobs/{id}/verification:
    patch:
      operationId: updateAdminJobVerification
      parameters:
        - $ref: "#/components/parameters/Id"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/JobVerificationUpdate"
      security:
        - ClerkCookieAuth: []
      responses:
        "200":
          description: Updated job verification
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AdminJob"
        "401":
          $ref: "#/components/responses/Error"
        "403":
          $ref: "#/components/responses/Error"
  /admin/jobs/{id}/repost:
    post:
      operationId: repostAdminJob
      security:
        - ClerkCookieAuth: []
      parameters:
        - $ref: "#/components/parameters/Id"
      responses:
        "201":
          description: Reposted job
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AdminJob"
        "401":
          $ref: "#/components/responses/Error"
        "403":
          $ref: "#/components/responses/Error"
  /admin/candidates:
    get:
      operationId: listAdminCandidates
      security:
        - ClerkCookieAuth: []
      parameters:
        - $ref: "#/components/parameters/Limit"
        - $ref: "#/components/parameters/Offset"
      responses:
        "200":
          description: Candidates
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AdminCandidatesPage"
        "401":
          $ref: "#/components/responses/Error"
        "403":
          $ref: "#/components/responses/Error"
  /admin/candidates/{id}/review:
    patch:
      operationId: updateAdminCandidateReview
      parameters:
        - $ref: "#/components/parameters/Id"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CandidateReviewUpdate"
      security:
        - ClerkCookieAuth: []
      responses:
        "200":
          description: Updated candidate review
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AdminCandidate"
        "401":
          $ref: "#/components/responses/Error"
        "403":
          $ref: "#/components/responses/Error"
  /admin/applications:
    get:
      operationId: listAdminApplications
      security:
        - ClerkCookieAuth: []
      parameters:
        - $ref: "#/components/parameters/Limit"
        - $ref: "#/components/parameters/Offset"
      responses:
        "200":
          description: Applications
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AdminApplicationsPage"
        "401":
          $ref: "#/components/responses/Error"
        "403":
          $ref: "#/components/responses/Error"
  /admin/applications/{id}:
    get:
      operationId: getAdminApplication
      security:
        - ClerkCookieAuth: []
      parameters:
        - $ref: "#/components/parameters/Id"
      responses:
        "200":
          description: Application detail
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AdminApplicationDetail"
        "401":
          $ref: "#/components/responses/Error"
        "403":
          $ref: "#/components/responses/Error"
  /admin/applications/{id}/status:
    patch:
      operationId: updateAdminApplicationStatus
      parameters:
        - $ref: "#/components/parameters/Id"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AdminApplicationStatusUpdate"
      security:
        - ClerkCookieAuth: []
      responses:
        "200":
          description: Updated application
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AdminApplication"
        "401":
          $ref: "#/components/responses/Error"
        "403":
          $ref: "#/components/responses/Error"
  /admin/audit-events:
    get:
      operationId: listAdminAuditEvents
      security:
        - ClerkCookieAuth: []
      parameters:
        - $ref: "#/components/parameters/Limit"
      responses:
        "200":
          description: Audit events
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AdminAuditEventsPage"
        "401":
          $ref: "#/components/responses/Error"
        "403":
          $ref: "#/components/responses/Error"
...
components:
  securitySchemes:
    ClerkCookieAuth:
      type: apiKey
      in: cookie
      name: __session
    ClerkBearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  parameters:
    Id:
      name: id
      in: path
      required: true
      schema:
        type: integer
    Limit:
      name: limit
      in: query
      schema:
        type: integer
        minimum: 1
        maximum: 100
        default: 25
    Offset:
      name: offset
      in: query
      schema:
        type: integer
        minimum: 0
        default: 0
  responses:
    Error:
      description: Error
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorEnvelope"
  schemas:
...
    ErrorEnvelope:
      type: object
      required:
        - error
      properties:
        error:
          type: string
    ApplicationStage:
      type: string
      enum:
        - Application Received
        - Reviewed
        - Shortlisted
        - Interview
        - Offered
        - Hired
        - Not Selected
        - New Applicant
        - AI Reviewed
        - Selected
        - Rejected
...
    ApplicationStatusHistory:
      type: object
      properties:
        id:
          type: integer
        status:
          $ref: "#/components/schemas/ApplicationStage"
        note:
          type:
            - string
            - "null"
        createdAt:
          type: string
          format: date-time
...
    AdminOverview:
      type: object
      required:
        - employers
        - jobs
        - candidates
        - applications
      properties:
        employers:
          type: integer
        jobs:
          type: integer
        candidates:
          type: integer
        applications:
          type: integer
    AdminFunctionCheck:
      type: object
      required:
        - api
        - database
        - counts
      properties:
        api:
          type: string
          enum:
            - ok
        database:
          type: string
          enum:
            - ok
        counts:
          $ref: "#/components/schemas/AdminOverview"
    AdminEmployer:
      type: object
      required:
        - id
        - companyName
        - contactName
        - verified
        - createdAt
      properties:
        id:
          type: integer
        companyName:
          type: string
        contactName:
          type: string
        verified:
          type: boolean
        createdAt:
          type: string
          format: date-time
    AdminJob:
      type: object
      required:
        - id
        - employerId
        - title
        - status
        - verificationStatus
        - createdAt
      properties:
        id:
          type: integer
        employerId:
          type: integer
        title:
          type: string
        department:
          type: string
        location:
          type: string
        status:
          type: string
          enum:
            - active
            - inactive
            - closed
        type:
          type: string
        expiresAt:
          type:
            - string
            - "null"
          format: date-time
        verificationStatus:
          type: string
          enum:
            - unverified
            - verified
            - needs_review
        createdAt:
          type: string
          format: date-time
    AdminCandidate:
      type: object
      required:
        - id
        - fullName
        - email
        - jobsMvReviewed
        - createdAt
      properties:
        id:
          type: integer
        fullName:
          type: string
        email:
          type: string
        phone:
          type:
            - string
            - "null"
        location:
          type:
            - string
            - "null"
        headline:
          type:
            - string
            - "null"
        jobsMvReviewed:
          type: boolean
        reviewedAt:
          type:
            - string
            - "null"
          format: date-time
        createdAt:
          type: string
          format: date-time
    CandidateReviewUpdate:
      type: object
      required:
        - jobsMvReviewed
      properties:
        jobsMvReviewed:
          type: boolean
        reviewNotes:
          type:
            - string
            - "null"
          maxLength: 5000
    AdminApplication:
      type: object
      required:
        - id
        - jobId
        - candidateId
        - status
        - createdAt
      properties:
        id:
          type: integer
        jobId:
          type: integer
        candidateId:
          type: integer
        status:
          $ref: "#/components/schemas/ApplicationStage"
        createdAt:
          type: string
          format: date-time
    AdminApplicationDetail:
      type: object
      required:
        - application
        - candidate
        - job
      properties:
        application:
          $ref: "#/components/schemas/AdminApplication"
        candidate:
          $ref: "#/components/schemas/AdminCandidate"
        job:
          $ref: "#/components/schemas/AdminJob"
        history:
          type: array
          items:
            $ref: "#/components/schemas/ApplicationStatusHistory"
    AdminEmployersPage:
      type: object
      required:
        - items
        - total
        - limit
        - offset
      properties:
        items:
          type: array
          items:
            $ref: "#/components/schemas/AdminEmployer"
        total:
          type: integer
        limit:
          type: integer
        offset:
          type: integer
    AdminJobsPage:
      type: object
      required:
        - items
        - total
        - limit
        - offset
      properties:
        items:
          type: array
          items:
            $ref: "#/components/schemas/AdminJob"
        total:
          type: integer
        limit:
          type: integer
        offset:
          type: integer
    AdminCandidatesPage:
      type: object
      required:
        - items
        - total
        - limit
        - offset
      properties:
        items:
          type: array
          items:
            $ref: "#/components/schemas/AdminCandidate"
        total:
          type: integer
        limit:
          type: integer
        offset:
          type: integer
    AdminApplicationsPage:
      type: object
      required:
        - items
        - total
        - limit
        - offset
      properties:
        items:
          type: array
          items:
            $ref: "#/components/schemas/AdminApplication"
        total:
          type: integer
        limit:
          type: integer
        offset:
          type: integer
    EmployerVerificationUpdate:
      type: object
      required:
        - verified
      properties:
        verified:
          type: boolean
    JobVerificationUpdate:
      type: object
      required:
        - verificationStatus
      properties:
        verificationStatus:
          type: string
          enum:
            - unverified
            - verified
            - needs_review
    AdminApplicationStatusUpdate:
      type: object
      required:
        - status
      properties:
        status:
          $ref: "#/components/schemas/ApplicationStage"
        note:
          type: string
          maxLength: 1000
    AdminAuditEvent:
      type: object
      required:
        - id
        - action
        - entityType
        - createdAt
      properties:
        id:
          type: integer
        action:
          type: string
        entityType:
          type: string
        entityId:
          type:
            - string
            - "null"
        metadata:
          type: object
          additionalProperties: true
        createdAt:
          type: string
          format: date-time
    AdminAuditEventsPage:
      type: object
      required:
        - items
      properties:
        items:
          type: array
          items:
            $ref: "#/components/schemas/AdminAuditEvent"
...